+++ student.rb
@@ -1,33 +1,14 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'person'
+
+class Student < Person
 
   def initialize(options = {})
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  include Phaseable
 
-  def set_phase(num)
-    response = ""
-    if num == @phase
-      response += "I'm doing phase #{@phase} again because "
-      response += "I put my learning first. I'm gonna rock it!"
-    else
-      response = "Oooh, phase #{num}. I hope I'm ready!"
-    end
-    @phase = num
-    response
-  end
+  include Teachable
 
-  def learn_stuff
-    response = ""
-    response += "WHOA! I've never thought of it quite like that before. "
-    response += "Now I feel like a genius!"
-    response
-  end
 end
